feat(Timeline/Hashtag): show confirmation sheet when muting
Shows a MuteHashtagConfirmationSheet when muting a hashtag, similar to
the MuteAccountConfirmationSheet. This also adds the option for the mute
to expire.
